# documentation/OpenAPI/de_interface.yaml
openapi: 3.0.3
info:
  title: "De Interface"
  description: API for frontend to interact with the backend
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: De interface
...
paths:
  /api/auth/register:
    post:
      summary: User registration
      description: Register a new user account
      operationId: registerUser
      tags:
        - "Login page"
      requestBody:
        description: Registration details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: "jona"
              email: "jona@student.uva.nl"
              password: "123456"
      responses:
        '200':
          description: Successful registration
          content:
            text/plain:
              schema:
                type: string
                description: |
                  JWT token containing user information.
                  When decoded, the payload contains: uuid (hex string), username, permission level, and expiration time.
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1dWlkIjoiMUYiLCJ1c2VybmFtZSI6ImJha2JlZXN0IiwicGVybWlzc2lvbiI6InVzZXIiLCJleHAiOjQ5MDI0fQ.signature"
        '400':
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterErrorResponse'
              examples:
                usernameError:
                  summary: Username validation error
                  value:
                    type: "username"
                    description: "Username too long (maximum is 40 characters)"
                usernameTaken:
                  summary: Username already exists
                  value:
                    type: "username"
                    description: "Username already in use"
                emailError:
                  summary: Email validation error
                  value:
                    type: "email"
                    description: "Invalid email format"
                passwordError:
                  summary: Password validation error
                  value:
                    type: "password"
                    description: "Password must be at least 6 characters long"
                otherError:
                  summary: Other server error
                  value:
                    type: "other"
                    description: "An unexpected error occurred"
  /api/auth/login:
    post:
      summary: User login
      description: Authenticate a user with username and password
      operationId: loginUser
      tags:
        - "Login page"
      requestBody:
        description: Login credentials
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: "jona"
              password: "123456"
      responses:
        '200':
          description: Successful authentication
          content:
            text/plain:
              schema:
                type: string
                description: |
                  JWT token containing user information.
                  When decoded, the payload contains: uuid (hex string), username, permission level, and expiration time.
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1dWlkIjoiMUYiLCJ1c2VybmFtZSI6ImJha2JlZXN0IiwicGVybWlzc2lvbiI6InVzZXIiLCJleHAiOjQ5MDI0fQ.signature"
        '400':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCredentials:
                  summary: Invalid username or password combination
                  value:
                    type: "invalid"
                    description: "Invalid username or password"
                usernameError:
                  summary: Username validation error
                  value:
                    type: "username"
                    description: "Username contains illegal characters"
                passwordError:
                  summary: Password validation error
                  value:
                    type: "password"
                    description: "Password must be at least 6 characters long"
                otherError:
                  summary: Other server error
                  value:
                    type: "other"
                    description: "An unexpected error occurred"
...
  /api/problem:
    get:
      summary: Get problem details
      description: Retrieve detailed information about a specific programming problem for the submission page
      operationId: getProblemDetails
      tags:
        - Submission page
      parameters:
        - in: query
          name: problem-id
          schema:
            type: integer
          required: true
          description: Database ID of the problem
          example: 1
      responses:
        '200':
          description: Successful retrieval of problem details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsResponse'
              example:
                problem-id: 1
                name: "Sort integer array"
                language: "c"
                difficulty: "easy"
                tags: ["array", "int"]
                short-description: "Lorem ipsum"
                long-description: "# Task\\nYou will be given an array of numbers."
                template-code: "int sort_odd(int n) {\\n    // Here comes your code.\\n}"
        '404':
          description: Problem not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemErrorResponse'
              example:
                error: "No problem found with id 1"
...
  /api/admin/add-problem:
    post:
      summary: Add new problem
      description: |
        Add a new programming problem to the system (admin only).

        **Authentication required**: Include JWT token in the Authorization header:
        `Authorization: Bearer [JWT]`
      operationId: AddProblem
      tags:
        - Admin page
      security:
        - bearerAuth: []
      requestBody:
        description: Problem details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProblemRequest'
            example:
              name: "Sort integer array"
              language: "c"
              difficulty: "easy"
              tags: ["array", "int"]
              short-description: "Lorem ipsum"
              long-description: "# Task\\nYou will be given an array of numbers."
              template-code: "int sort_odd(int n) {\\n    // Here comes your code.\\n}"
      responses:
        '200':
          description: Problem successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddProblemResponse'
              example:
                problem-id: "5"
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDetailedErrorResponse'
              example:
                error: "Validation failed"
                description: "Title is required\\nDifficulty must be one of: easy, medium, hard"
        '401':
          description: Unauthorized - user is not admin or JWT invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDetailedErrorResponse'
              example:
                error: "Unauthorized"
                description: "User does not have admin permissions"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDetailedErrorResponse'

  ### ───────────────────────── Problems list (overview page) ─────────────────────────
...
components:
  schemas:
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          description: Desired username
          example: "jona"
        email:
          type: string
          format: email
          description: User's email address
          example: "jona@student.uva.nl"
        password:
          type: string
          format: password
          description: User's password
          example: "123456"
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: User's username
          example: "jona"
        password:
          type: string
          format: password
          description: User's password
          example: "123456"
    ErrorResponse:
      type: object
      required:
        - type
        - description
      properties:
        type:
          type: string
          description: Error type indicating which field caused the error
          enum:
            - invalid
            - username
            - email
            - password
            - other
          example: "invalid"
        description:
          type: string
          description: Human-readable error message to display to the user (starts with capital letter, no period)
          example: "Invalid username or password"
    RegisterErrorResponse:
      type: object
      required:
        - type
        - description
      properties:
        type:
          type: string
          description: Error type indicating which field caused the error
          enum:
            - username
            - email
            - password
            - other
          example: "username"
        description:
          type: string
          description: Human-readable error message to display to the user (starts with capital letter, no period)
          example: "Username too long (maximum is 40 characters)"
...
    ProblemDetailsResponse:
      type: object
      required:
        - problem-id
        - name
        - language
        - difficulty
        - tags
        - short-description
        - long-description
        - template-code
      properties:
        problem-id:
          type: integer
          description: Database ID of the problem
          example: 1
        name:
          type: string
          description: Name of the programming problem
          example: "Sort integer array"
        language:
          type: string
          description: Programming language for the problem
          example: "c"
        difficulty:
          type: string
          description: Difficulty level of the problem
          example: "easy"
        tags:
          type: array
          description: List of tags associated with the problem
          items:
            type: string
          example: ["array", "int"]
        short-description:
          type: string
          description: Brief description of the problem
          example: "Lorem ipsum"
        long-description:
          type: string
          description: Detailed problem description in Markdown format
          example: "# Task\\nYou will be given an array of numbers."
        template-code:
          type: string
          description: Template code to start with
          example: "int sort_odd(int n) {\\n    // Here comes your code.\\n}"
    ProblemErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "No problem found with id 1"
...
    AdminDetailedErrorResponse:
      type: object
      required:
        - error
        - description
      properties:
        error:
          type: string
          description: Short error description
          example: "Validation failed"
        description:
          type: string
          description: Detailed error explanation, multiple errors separated by newlines
          example: "Title is required\\nDifficulty must be one of: easy, medium, hard"
    AddProblemRequest:
      type: object
      required:
        - name
        - language
        - difficulty
        - tags
        - short-description
        - long-description
        - template-code
      properties:
        name:
          type: string
          description: Problem name (must be unique)
          example: "Sort integer array"
        language:
          type: string
          description: Programming language
          example: "c"
        difficulty:
          type: string
          description: Problem difficulty level
          enum:
            - easy
            - medium
            - hard
          example: "easy"
        tags:
          type: array
          description: Problem tags
          items:
            type: string
          example: ["array", "int"]
        short-description:
          type: string
          description: Brief problem description
          example: "Lorem ipsum"
        long-description:
          type: string
          description: Detailed problem description in Markdown format
          example: "# Task\\nYou will be given an array of numbers."
        template-code:
          type: string
          description: Template code for the problem
          example: "int sort_odd(int n) {\\n    // Here comes your code.\\n}"
    AddProblemResponse:
      type: object
      required:
        - problem-id
      properties:
        problem-id:
          type: string
          description: Database ID of the created problem
          example: "5"

    # ───────────────────────── Problems list ─────────────────────────
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from login or register endpoints.

        **How to use**: Include the token in the Authorization header:
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        ```

        The JWT payload contains:
        - uuid: User's unique identifier (hexadecimal string)
        - username: User's username
        - permission: Permission level ("user" or "admin")
        - exp: Token expiration time in seconds

        Example payload:
        {
          "uuid": "1F",
          "username": "bakbeest",
          "permission": "user",
          "exp": 49024
        }